% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actuarialmetrics-package.R
\docType{package}
\name{actuarialmetrics-package}
\alias{actuarialmetrics}
\alias{actuarialmetrics-package}
\title{actuarialmetrics: Actuarial science tools for R}
\description{
Package that builds actuarial science tools with R. The pacakge offer simple visualisation for classic model evaluation tools such as double lift charts, loss ratios lift charts or gini index curves.
}
\author{
\strong{Maintainer}: Olivier Côté \email{olivier.cote.12@ulaval.ca}

}
\keyword{internal}
