% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lr_table.R
\name{get_lr_table}
\alias{get_lr_table}
\title{Prepare a dataset for actuarial tools}
\usage{
get_lr_table(
  pred_table,
  ref_name,
  comp_name,
  expo_name = NULL,
  loss_name = NULL,
  n_cuts = 10,
  balance = TRUE
)
}
\arguments{
\item{pred_table}{A \code{dataframe} containing a column per ratemaking structure (at least 2, \code{ref_name} and \code{comp_name}), a column for exposure (\code{expo_name}, which is optionnal) and a column for the response variable \code{loss_name}.}

\item{ref_name}{The name of the reference ratemaking structure. It is the name of a column of \code{pred_table}.}

\item{comp_name}{The name of the ratemaking structure to compare. It is the name of a column of \code{pred_table}.}

\item{expo_name}{(\emph{optionnal}) The name of the loss exposure for each observation. It is the name of a column of \code{pred_table}. If nothing is specified, constant exposure of 1 is assumed accross all observations.}

\item{loss_name}{The name of the response variable for each observation. It is the name of a column of \code{pred_table}.}

\item{n_cuts}{The amount of groups for the data aggregation. The data will be sorted by relativity and grouped (equal exposure groups) by ascending relativity order. Default value is 10.}

\item{balance}{boolean indicating wether or not to balance total premiums with the total response variable. The default value is \code{TRUE} since the objective of the analysis is to assess segmentation.}
}
\value{
A table ready to input to \emph{double lift chart} or \emph{Loss ratio lift} functions from \code{actuarialmetrics}.
}
\description{
Prepare a dataset for actuarial tools
}
\examples{
library(actuarialmetrics)
data("pred_table")
table_to_g <- get_lr_table(pred_table,
                           ref_name = "mod2",
                             comp_name = "mod3",
                             loss_name = "Y",
                             n_cuts = 6)

}
